+++ tests/Unit/CachedBuilderTest.php
@@ -10,6 +10,7 @@ use GeneaLabs\LaravelModelCaching\Tests\Fixtures\UncachedBook;
 use GeneaLabs\LaravelModelCaching\Tests\Fixtures\UncachedProfile;
 use GeneaLabs\LaravelModelCaching\Tests\Fixtures\UncachedPublisher;
 use GeneaLabs\LaravelModelCaching\Tests\Fixtures\UncachedStore;
+use GeneaLabs\LaravelModelCaching\Tests\Fixtures\Http\Resources\Author as AuthorResource;
 use GeneaLabs\LaravelModelCaching\Tests\TestCase;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 
@@ -53,7 +54,7 @@ class CachedBuilderTest extends TestCase
                 'genealabslaravelmodelcachingtestsfixturesauthor',
                 'genealabslaravelmodelcachingtestsfixturesbook'
             ])
-            ->get('genealabslaravelmodelcachingtestsfixturesauthor_1_2_3_4_5_6_7_8_9_10-genealabslaravelmodelcachingtestsfixturesbooks');
+            ->get('genealabslaravelmodelcachingtestsfixturesauthor-books');
 
         $this->assertNull($results);
     }
@@ -436,6 +437,36 @@ class CachedBuilderTest extends TestCase
         );
 
         $this->assertEmpty($authors->diffAssoc($cachedResults));
+        $this->assertEmpty($liveResults->diffAssoc($cachedResults));
+    }
+
+    public function testLazyLoadedRelationshipResolvesThroughCachedBuilder()
+    {
+        $books = (new Author)->first()->books;
+        $key = 'genealabslaravelmodelcachingtestsfixturesbook-books.author_id_1';
+        $tags = [
+            'genealabslaravelmodelcachingtestsfixturesbook',
+        ];
+
+        $cachedResults = cache()->tags($tags)->get($key);
+        $liveResults = (new UncachedAuthor)->first()->books;
+
+        $this->assertEmpty($books->diffAssoc($cachedResults));
+        $this->assertEmpty($liveResults->diffAssoc($cachedResults));
+    }
+
+    public function testLazyLoadingOnResourceIsCached()
+    {
+        $books = (new AuthorResource((new Author)->first()))->books;
+        $key = 'genealabslaravelmodelcachingtestsfixturesbook-books.author_id_1';
+        $tags = [
+            'genealabslaravelmodelcachingtestsfixturesbook',
+        ];
+
+        $cachedResults = cache()->tags($tags)->get($key);
+        $liveResults = (new UncachedAuthor)->first()->books;
+
+        $this->assertEmpty($books->diffAssoc($cachedResults));
         $this->assertEmpty($liveResults->diffAssoc($cachedResults));
     }
 }